// resources/js/types/auth.ts
import type { Site } from './paklaring';

export type User = {
    id: number;
    name: string;
    email: string;
    avatar?: string;
    role: 'super_admin' | 'site_admin';
    site_id: number | null;
    site?: Site | null;
    email_verified_at: string | null;
    two_factor_enabled?: boolean;
    created_at: string;
    updated_at: string;
    [key: string]: unknown;
};

export type Auth = {
    user: User;
};

/* @chisel-passkeys */
export type Passkey = {
    id: number;
    name: string;
    authenticator: string | null;
    created_at_diff: string;
    last_used_at_diff: string | null;
};
/* @end-chisel-passkeys */

export type TwoFactorSetupData = {
    svg: string;
    url: string;
};

export type TwoFactorSecretKey = {
    secretKey: string;
};